+++ kernel/drivers/ipc/xddp.c
@@ -105,7 +105,7 @@ static struct xnpnode_link __xddp_pnode = {
 #endif /* !CONFIG_XENO_OPT_VFILE */
 
 static void __xddp_flush_pool(struct xnheap *heap,
-			      void *poolmem, u_long poolsz, void *cookie)
+			      void *poolmem, u32 poolsz, void *cookie)
 {
 	free_pages_exact(poolmem, poolsz);
 }
